requirements.txt
# =========================
# CORE
# =========================
fastapi==0.128.7
uvicorn[standard]==0.40.0

# =========================
# DATABASE (ASYNC)
# =========================
SQLAlchemy==2.0.46
asyncpg==0.31.0

# =========================
# AUTH / SECURITY
# =========================
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# =========================
# TELEGRAM
# =========================
aiogram==3.24.0

# =========================
# HTTP CLIENT (LLM / APIs)
# =========================
httpx==0.28.1

# =========================
# UTILITIES
# =========================
python-dateutil==2.9.0.post0
dateparser==1.3.0

# =========================
# ENV / CONFIG
# =========================
pydantic-settings==2.12.0
python-dotenv==1.2.1

# =========================
# DATABASE MIGRATIONS
# =========================
alembic==1.18.4

openai==2.21.0

dateparser==1.3.0